[SPARK-31893][ML] Add a generic ClassificationSummary trait
